fix: refactoring bugs 🐛
