+++ bot/src/main/kotlin/me/melijn/bot/commands/music/MusicExtension.kt
@@ -1,4 +1,4 @@
-package me.melijn.bot.commands
+package me.melijn.bot.commands.music
 
 import com.kotlindiscord.kord.extensions.checks.guildFor
 import com.kotlindiscord.kord.extensions.commands.Arguments
@@ -10,18 +10,12 @@ import com.kotlindiscord.kord.extensions.commands.converters.impl.*
 import com.kotlindiscord.kord.extensions.extensions.Extension
 import com.kotlindiscord.kord.extensions.types.editingPaginator
 import com.kotlindiscord.kord.extensions.types.respond
-import com.kotlindiscord.kord.extensions.utils.suggestStringMap
 import dev.kord.common.entity.ChannelType
 import dev.kord.rest.builder.message.create.embed
 import dev.schlaubi.lavakord.audio.Link
 import dev.schlaubi.lavakord.kord.connectAudio
-import kotlinx.datetime.UtcOffset
-import kotlinx.datetime.toInstant
-import kotlinx.datetime.toJavaInstant
 import me.melijn.apkordex.command.KordExtension
 import me.melijn.bot.Melijn
-import me.melijn.bot.database.manager.PlaylistManager
-import me.melijn.bot.database.manager.PlaylistTrackManager
 import me.melijn.bot.model.PartialUser
 import me.melijn.bot.model.TrackSource
 import me.melijn.bot.music.*
@@ -33,8 +27,6 @@ import me.melijn.bot.utils.KordExUtils.userIsOwner
 import me.melijn.bot.utils.StringsUtil.ansiFormat
 import me.melijn.bot.utils.TimeUtil.formatElapsed
 import me.melijn.bot.utils.intRanges
-import me.melijn.bot.utils.optionalIntRanges
-import me.melijn.bot.utils.playlist
 import me.melijn.bot.utils.shortTime
 import me.melijn.bot.web.api.WebManager
 import me.melijn.kordkommons.utils.StringUtils
@@ -52,311 +44,40 @@ class MusicExtension : Extension() {
     private val webManager by inject<WebManager>()
     private val trackLoader by inject<TrackLoader>()
 
-    inner class FollowUserArgs : Arguments() {
-        val target = optionalUser {
-            name = "target"
-            description = "MusicPlayer will follow spotify status"
-        }
-    }
-
-    inner class SeekArgs : Arguments() {
-        val time = shortTime {
-            name = "timeStamp"
-            description = "format mm:ss or hh:mm:ss (e.g. 1:35 for 1 minute 35 seconds)"
-        }
-    }
-
-    inner class MoveArgs : Arguments() {
-        val from = int {
-            name = "from"
-            description = "Index of the track you want to move (see queue command for viewing indexes)"
-            validate {
-                failIfInvalidTrackIndex(value)
+    companion object {
+        suspend fun ValidationContext<*>.failIfInvalidTrackIndex(
+            index: Int?,
+            trackManagerFun: suspend ValidationContext<*>.() -> TrackManager = {
+                context.getGuild()!!.asGuild().getTrackManager()
+            }
+        ) {
+            val noVarMoment = index ?: return
+            failIf(context.tr("move.invalidIndex", noVarMoment)) {
+                val trackManager = trackManagerFun()
+                index > trackManager.queue.size || index < 1
             }
         }
-        val to = int {
-            name = "to"
-            description = "Index where the track needs to be moved to"
-            validate {
-                failIfInvalidTrackIndex(value)
-            }
-        }
-    }
 
-    inner class RemoveArgs : Arguments() {
-        val positions = intRanges {
-            name = "positions"
-            description = "Number ranges or numbers seperated by commas (e.g. 1,5,9-15)"
-        }
-    }
-
-    inner class PlaylistAddArgs : Arguments() {
-        val playlist = string {
-            name = "playlist"
-            description = "New or existing playlist name or existing playlist index"
-            autoComplete {
-                val playlists = playlistManager.getByIndex1(user.id.value)
-                suggestStringMap(playlists.associate { it.name to it.name })
-            }
-        }
-        val trackIndexes = optionalIntRanges {
-            name = "trackindex"
-            description = "Index of a track in queue, see /queue"
-            validate {
-                val varmoment = value ?: return@validate
-                val trackManager = context.getGuild()!!.asGuild().getTrackManager()
-                varmoment.list.forEach {
-                    for (i in it) failIfInvalidTrackIndex(i) { trackManager }
-                }
-            }
-        }
-    }
-
-    inner class PlaylistRemoveArgs : Arguments() {
-        val playlist = playlist {
-            name = "playlist"
-            description = "Existing playlist name"
-
-            autoComplete {
-                val playlists = playlistManager.getByIndex1(user.id.value)
-                suggestStringMap(playlists.associate { it.name to it.name })
-            }
-        }
-        val trackIndexes = intRanges {
-            name = "trackindex"
-            description = "Index of a track in the playlist, see `/playlist tracks`"
-            validate {
-                val trackCount = playlistTrackManager.getTrackCount(playlist.parsed)
-                value.list.any {
-                    val from = it.first
-                    val to = it.last
-                    failIf(
-                        context.tr(
-                            "playlist.invalidTrackIndex",
-                            from == to, from.toString(), to.toString()
-                        )
-                    ) {
-                        from >= trackCount || from < 0 || to >= trackCount || to < 0
+        /**
+         * Bot tries to join the user's vc if it isn't in a voiceChannel, if user is also not in vc, respond with error
+         * @return true for failure
+         */
+        suspend fun PublicSlashCommandContext<*>.tryJoinUser(link: Link): Boolean {
+            if (link.state != Link.State.CONNECTED) {
+                val vc = member?.getVoiceStateOrNull()?.channelId
+                if (vc == null) {
+                    respond {
+                        content = tr("music.userNotInVC")
                     }
+                    return true
                 }
+                link.connectAudio(vc.value)
             }
+            return false
         }
     }
-
-    inner class PlaylistListArgs : Arguments() {
-        val playlist = playlist {
-            name = "playlist"
-            description = "Existing playlist name"
-
-            autoComplete {
-                val playlists = playlistManager.getByIndex1(user.id.value)
-                suggestStringMap(playlists.associate { it.name to it.name })
-            }
-        }
-    }
-
-    private suspend fun ValidationContext<*>.failIfInvalidTrackIndex(
-        index: Int?,
-        trackManagerFun: suspend ValidationContext<*>.() -> TrackManager = {
-            context.getGuild()!!.asGuild().getTrackManager()
-        }
-    ) {
-        val noVarMoment = index ?: return
-        failIf(context.tr("move.invalidIndex", noVarMoment)) {
-            val trackManager = trackManagerFun()
-            index > trackManager.queue.size || index < 1
-        }
-    }
-
-    val playlistManager by inject<PlaylistManager>()
-    val playlistTrackManager by inject<PlaylistTrackManager>()
 
     override suspend fun setup() {
-        publicGuildSlashCommand {
-            name = "playlist"
-            description = "Manage playlist things"
-
-
-            publicSubCommand(::PlaylistListArgs) {
-                name = "load"
-                description = "Loads all tracks of the playlist into the queue"
-
-                action {
-                    val playlist = this.arguments.playlist.parsed
-                    val requester = PartialUser.fromKordUser(user.asUser())
-                    val tracks = playlistTrackManager.getMelijnTracksInPlaylist(playlist, requester)
-                    val guild = this.guild!!.asGuild()
-                    val trackManager = guild.getTrackManager()
-                    val link = trackManager.link
-                    if (tryJoinUser(link)) return@action
-
-                    for (track in tracks) trackManager.queue(track, QueuePosition.BOTTOM)
-
-                    respond {
-                        content = tr("playlist.load.queued", tracks.size, playlist.name.escapeMarkdown())
-                    }
-                }
-            }
-
-            publicSubCommand(::PlaylistAddArgs) {
-                name = "add"
-                description = "Adds a track to your playlist"
-
-                action {
-                    val guild = guild!!.asGuild()
-                    val playlistName = arguments.playlist.parsed
-                    val trackManager = guild.getTrackManager()
-                    val playingTrack = trackManager.playingTrack
-
-                    // intRanges into tracks conversion
-                    val hashSet = HashSet<Int>()
-                    val trackIndexes = arguments.trackIndexes.parsed
-                    val targetTracks = trackIndexes?.list?.let { ranges ->
-                        for (range in ranges)
-                            for (i in range) hashSet.add(i-1)
-
-                        val shouldContainPlayingTrack = hashSet.remove(-1)
-                        val tracks = trackManager.getTracksByIndexes(hashSet).toMutableList()
-                        if (shouldContainPlayingTrack) playingTrack?.let { tracks.add(it) }
-                        tracks
-                    } ?: buildList { playingTrack?.let { add(it) } }
-
-                    // potentially create non-existant playlist
-                    val existingPlaylist = playlistManager.getByNameOrDefault(user.id, playlistName)
-                    playlistManager.store(existingPlaylist)
-
-                    val oldTrackCount = playlistTrackManager.getTrackCount(existingPlaylist)
-
-                    // add each track to the existing playlist
-                    for (track in targetTracks)
-                        playlistTrackManager.newTrack(existingPlaylist, track)
-
-                    if (targetTracks.size == 1) {
-                        val targetTrack = targetTracks.first()
-                        respond {
-                            content = tr(
-                                "playlist.add.added", targetTrack.title.escapeMarkdown(),
-                                playlistName.escapeMarkdown(), oldTrackCount
-                            )
-                        }
-                    } else {
-                        respond {
-                            content = tr(
-                                "playlist.add.addedMany",
-                                targetTracks.size,
-                                playlistName.escapeMarkdown(),
-                                oldTrackCount,
-                                oldTrackCount + targetTracks.size - 1
-                            )
-                        }
-                    }
-                }
-            }
-
-            publicSubCommand(::PlaylistRemoveArgs) {
-                name = "remove"
-                description = "Removes tracks from your playlist"
-
-                action {
-                    val existingPlaylist = arguments.playlist.parsed
-                    val tracks = playlistTrackManager.getTracksInPlaylist(existingPlaylist)
-                    val toRemove = tracks.withIndex().filter { (i, _) ->
-                        arguments.trackIndexes.parsed.list.any { it.contains(i) }
-                    }
-                    if (toRemove.size == 1) {
-                        val (index, track) = toRemove[0]
-                        playlistTrackManager.delete(track)
-                        respond {
-                            content = tr(
-                                "playlist.remove.removedTrack",
-                                index,
-                                track.url,
-                                track.title.escapeMarkdown(),
-                                existingPlaylist.name.escapeMarkdown()
-                            )
-                        }
-                    } else {
-                        playlistTrackManager.deleteAll(toRemove.map { it.value })
-                        respond {
-                            content = tr(
-                                "playlist.remove.removedTracks",
-                                toRemove.size,
-                                existingPlaylist.name.escapeMarkdown()
-                            )
-                        }
-                    }
-
-                    if (toRemove.size == tracks.size) playlistManager.delete(existingPlaylist)
-                }
-            }
-
-            publicSubCommand {
-                name = "playlists"
-                description = "Lists all your playlists"
-
-                action {
-                    val existingPlaylist = playlistManager.getPlaylistsOfUserWithTrackCount(user.id)
-
-                    var description = "```INI\n# index - public - [name] - tracks - [created]"
-
-                    existingPlaylist.entries.withIndex().forEach { (index, playlistWithCount) ->
-                        val (playlist, count) = playlistWithCount
-                        description += "\n" + tr(
-                            "playlist.list.all.playlistEntry", index, playlist.public, playlist.name, count,
-                            java.util.Date.from(playlist.created.toInstant(UtcOffset.ZERO).toJavaInstant())
-                        )
-                    }
-                    description += "```"
-
-                    editingPaginator {
-                        val parts = StringUtils.splitMessageWithCodeBlocks(description)
-                        for (part in parts) {
-                            page {
-                                this.title = tr("playlist.list.all.listTitle", user.asUser().tag)
-                                this.description = part
-                            }
-                        }
-                    }.send()
-                }
-            }
-
-            publicSubCommand(::PlaylistListArgs) {
-                name = "tracks"
-                description = "Lists your playlist tracks"
-
-                action {
-                    val existingPlaylist = arguments.playlist.parsed
-                    val tracks = playlistTrackManager.getTracksInPlaylist(existingPlaylist)
-                    var totalDuration = Duration.ZERO
-                    var description = ""
-
-                    tracks.withIndex().forEach { (index, trackData) ->
-                        totalDuration += trackData.length
-                        description += "\n" + tr(
-                            "queue.queueEntry", index, trackData.url, trackData.title,
-                            trackData.length.formatElapsed()
-                        )
-                    }
-
-                    description += tr(
-                        "playlist.list.fakeFooter",
-                        totalDuration.formatElapsed(),
-                        tracks.size
-                    )
-
-                    editingPaginator {
-                        val parts = StringUtils.splitMessage(description)
-                        for (part in parts) {
-                            page {
-                                this.title = tr("playlist.list.listTitle", user.asUser().tag)
-                                this.description = part
-                            }
-                        }
-                    }.send()
-                }
-            }
-        }
-
         publicGuildSlashCommand {
             name = "shuffle"
             description = "shuffles the queue once"
@@ -409,10 +130,12 @@ class MusicExtension : Extension() {
                 val from = arguments.positions.parsed
                 val guild = guild!!.asGuild()
                 val trackManager = guild.getTrackManager()
+
+                // collect tracks from queue to remove from intRanges
                 val toRemove = mutableListOf<Track>()
                 for (range in from.list)
                     for (i in range) {
-                        val element = trackManager.queue.get(i - 1)
+                        val element = trackManager.queue.get(i - 1) // queue is offset by one
                         if (!toRemove.contains(element))
                             toRemove.add(element)
                     }
@@ -705,7 +428,7 @@ class MusicExtension : Extension() {
             }
         }
 
-        publicGuildSlashCommand(::VCArgs) {
+        publicGuildSlashCommand(MusicExtension::VCArgs) {
             name = "summon"
             description = "bot joins your channel"
 
@@ -726,7 +449,7 @@ class MusicExtension : Extension() {
             }
         }
 
-        publicGuildSlashCommand(::PlayArgs) {
+        publicGuildSlashCommand(MusicExtension::PlayArgs) {
             name = "play"
             description = "bot joins your channel and plays moosic"
 
@@ -828,22 +551,43 @@ class MusicExtension : Extension() {
         }
     }
 
-    /**
-     * Bot tries to join the user's vc if it isn't in a voiceChannel, if user is also not in vc, respond with error
-     * @return true for failure
-     */
-    private suspend fun PublicSlashCommandContext<*>.tryJoinUser(link: Link): Boolean {
-        if (link.state != Link.State.CONNECTED) {
-            val vc = member?.getVoiceStateOrNull()?.channelId
-            if (vc == null) {
-                respond {
-                    content = tr("music.userNotInVC")
-                }
-                return true
-            }
-            link.connectAudio(vc.value)
+
+    inner class FollowUserArgs : Arguments() {
+        val target = optionalUser {
+            name = "target"
+            description = "MusicPlayer will follow spotify status"
         }
-        return false
+    }
+
+    inner class SeekArgs : Arguments() {
+        val time = shortTime {
+            name = "timeStamp"
+            description = "format mm:ss or hh:mm:ss (e.g. 1:35 for 1 minute 35 seconds)"
+        }
+    }
+
+    inner class MoveArgs : Arguments() {
+        val from = int {
+            name = "from"
+            description = "Index of the track you want to move (see queue command for viewing indexes)"
+            validate {
+                failIfInvalidTrackIndex(value)
+            }
+        }
+        val to = int {
+            name = "to"
+            description = "Index where the track needs to be moved to"
+            validate {
+                failIfInvalidTrackIndex(value)
+            }
+        }
+    }
+
+    inner class RemoveArgs : Arguments() {
+        val positions = intRanges {
+            name = "positions"
+            description = "Number ranges or numbers seperated by commas (e.g. 1,5,9-15)"
+        }
     }
 
     private class PlayArgs : Arguments() {
