+++ bot/src/main/kotlin/me/melijn/bot/database/manager/PlaylistTrackManager.kt
@@ -7,8 +7,13 @@ import me.melijn.bot.database.model.PlaylistSpotifyTrack
 import me.melijn.bot.database.model.PlaylistTrack
 import me.melijn.bot.model.PartialUser
 import me.melijn.bot.music.*
+import me.melijn.bot.utils.KoinUtil.inject
 import me.melijn.gen.PlaylistData
+import me.melijn.gen.PlaylistFetchedTrackData
+import me.melijn.gen.PlaylistSpotifyTrackData
 import me.melijn.gen.PlaylistTrackData
+import me.melijn.gen.database.manager.AbstractPlaylistFetchedTrackManager
+import me.melijn.gen.database.manager.AbstractPlaylistSpotifyTrackManager
 import me.melijn.gen.database.manager.AbstractPlaylistTrackManager
 import me.melijn.kordkommons.database.DriverManager
 import me.melijn.kordkommons.utils.TimeUtil
@@ -18,11 +23,15 @@ import java.util.*
 @Inject
 class PlaylistTrackManager(override val driverManager: DriverManager) : AbstractPlaylistTrackManager(driverManager) {
 
+    private val playlistFetchedTrackManager by inject<PlaylistFetchedTrackManager>()
+    private val playlistSpotifyTrackManager by inject<PlaylistSpotifyTrackManager>()
+
     fun newTrack(playlist: PlaylistData, track: Track) {
+        val trackId = UUID.randomUUID()
         store(
             PlaylistTrackData(
                 playlist.playlistId,
-                UUID.randomUUID(),
+                trackId,
                 track.title,
                 track.url,
                 track.isStream,
@@ -31,6 +40,21 @@ class PlaylistTrackManager(override val driverManager: DriverManager) : Abstract
                 TimeUtil.localDateTimeNow()
             )
         )
+        when (track) {
+            is SpotifyTrack -> playlistSpotifyTrackManager.store(
+                PlaylistSpotifyTrackData(trackId, track.author, track.identifier)
+            )
+            is FetchedTrack -> playlistFetchedTrackManager.store(
+                PlaylistFetchedTrackData(
+                    trackId,
+                    track.track,
+                    track.author,
+                    track.identifier,
+                    track.sourceType,
+                    track.trackInfoVersion.toInt()
+                )
+            )
+        }
     }
 
     fun getTracksInPlaylist(playlist: PlaylistData): List<PlaylistTrackData> {
@@ -44,8 +68,8 @@ class PlaylistTrackManager(override val driverManager: DriverManager) : Abstract
     }
 
     fun getMelijnTracksInPlaylist(playlist: PlaylistData, requester: PartialUser): List<Track> {
-        return scopedTransaction {
-            val sortableTracks = mutableListOf<Triple<UUID, LocalDateTime, Track>>()
+        val sortableTracks = mutableListOf<Triple<UUID, LocalDateTime, Track>>()
+        scopedTransaction {
             PlaylistTrack.join(PlaylistFetchedTrack, JoinType.INNER) {
                 PlaylistTrack.trackType.eq(TrackType.FETCHED)
                     .and(PlaylistTrack.trackId.eq(PlaylistFetchedTrack.trackId))
@@ -68,10 +92,12 @@ class PlaylistTrackManager(override val driverManager: DriverManager) : Abstract
                     )
                 )
             }
-            PlaylistTrack.join(PlaylistFetchedTrack, JoinType.INNER) {
-                PlaylistTrack.trackType.eq(TrackType.SPOTIFY)
-                    .and(PlaylistTrack.trackId.eq(PlaylistFetchedTrack.trackId))
-            }.selectAll().forEach {
+            PlaylistTrack.join(PlaylistSpotifyTrack, JoinType.INNER) {
+                PlaylistTrack.trackId.eq(PlaylistSpotifyTrack.trackId)
+            }.select {
+                PlaylistTrack.playlistId.eq(playlist.playlistId)
+                    .and(PlaylistTrack.trackType.eq(TrackType.SPOTIFY))
+            }.forEach {
                 sortableTracks.add(
                     Triple(
                         it[PlaylistTrack.trackId], it[PlaylistTrack.addedTime], SpotifyTrack(
@@ -80,16 +106,15 @@ class PlaylistTrackManager(override val driverManager: DriverManager) : Abstract
                             it[PlaylistTrack.url],
                             it[PlaylistSpotifyTrack.identifier],
                             it[PlaylistTrack.isStream],
-                            TrackData.fromNow(requester, it[PlaylistFetchedTrack.identifier]),
+                            TrackData.fromNow(requester, it[PlaylistSpotifyTrack.identifier]),
                             it[PlaylistTrack.length]
                         )
                     )
                 )
             }
-            sortableTracks.sortedBy { it.first }.sortedBy { it.second }.map { it.third }
         }
+        return sortableTracks.sortedBy { it.first }.sortedBy { it.second }.map { it.third }
     }
-
 
     fun deleteAll(tracks: List<PlaylistTrackData>) {
         scopedTransaction {
@@ -106,3 +131,9 @@ class PlaylistTrackManager(override val driverManager: DriverManager) : Abstract
         }
     }
 }
+
+@Inject
+class PlaylistSpotifyTrackManager(driverManager: DriverManager) : AbstractPlaylistSpotifyTrackManager(driverManager)
+
+@Inject
+class PlaylistFetchedTrackManager(driverManager: DriverManager) : AbstractPlaylistFetchedTrackManager(driverManager)
