+++ bot/src/main/resources/translations/melijn/strings.properties
@@ -76,12 +76,12 @@ loop.single.looped=Track looping **{0, select, true {enabled} other {disabled}}*
 pause.paused=**{0, select, true {Paused} other {Resumed}}** the playing track
 shuffle.shuffled=Shuffled the queue
 
+
 playlist.add.added=Added **{0}** to your playlist `{1}` at #{2}
-playlist.add.addedMany=Added **{0}** tracks to your playlist `{1}` at positions #{2} -> ${3}
+playlist.add.addedMany=Added **{0}** tracks to your playlist `{1}` at positions #{2} -> #{3}
 playlist.list.listTitle=Playlist of {0}
 playlist.list.fakeFooter=\n\nPlaylist duration: **{0}**\
   \nPlaylist size: **{1}**
-playlist.playlistArgDoesNotExist=`{0}` is not a valid playlist, it does not exist
 playlist.list.all.playlistEntry={0} - {1} - [{2}] - {3} - [{4, date, short} {4, time, short}]
 playlist.list.all.listTitle=Playlist List of {0}
 playlist.invalidTrackIndex=`{0, select, true {{1}} other {{1}-{2}}}` is not a valid playlist-track index\
